requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
paramiko==3.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # ← FIX: Downgrade from 5.0.0 to 4.0.1
cryptography==41.0.7
pydantic==2.5.0
python-multipart==0.0.6
aiosmtplib==2.0.1  # ← FIX: Downgrade from 5.0.0 to 2.4.0
python-dotenv==1.0.0
aiofiles==23.2.1